fix typescript circular dep issue
