+++ lib/types/index.ts
@@ -22,5 +22,6 @@ export * from './OAuth';
 export * from './OktaAuthOptions';
 export * from './Storage';
 export * from './Token';
+export * from './TokenManager';
 export * from './UserClaims';
 export * from './AuthState';
